[OPP-1422] refaktorere sakerRouting ut i dyplenker
